Update requests of device flow for public client
This commit makes a change in OIDC Device Flow to include the client_id
in the request body without performing basic authentication when no
client_secret is specified.

According to RFC8628, in the Device Authorization Request [1] and Device
Access Token Request [2], if the client is not authenticating with the
authorization server, the client_id should be included in the request
body.

In addition, this commit adds missing unit tests for the
OidcDeviceAuthorization.

[1] https://datatracker.ietf.org/doc/html/rfc8628#section-3.1
[2] https://datatracker.ietf.org/doc/html/rfc8628#section-3.4

Closes-bug: #2053965
Change-Id: Ieada594dc68e152e176969258f098f7d064801f4